Move One Step A*
Changed a few functions to return more values so that the correct result can be outputted.

- calcClosestTarget() returns a 2-tuple of (distance, target)
- findPath() returns a list of 3-tuples now instead of 2-tuples - (x,y,moveType)
- calcStates now returns a list of 3-tuples - (x,y,moveType)
- moveType is either "SWING" or "SLIDE"

Tokens now move once and then their path is calculated again. If there is no path available, move the token in any random direction calculated from calcStates(). This keeps happening until no targets remain for every token.

Also changed calcStates() to one for loop just to make it easier to modify in the future